chore: smaller components for Credit Class Details
